Split statistics routines from math.cc
