+++ chainerx_cc/chainerx/cuda/cuda_device/reduction.cu
@@ -132,64 +132,6 @@ public:
 
 CHAINERX_CUDA_REGISTER_KERNEL(SumKernel, CudaSumKernel);
 
-template <typename T>
-struct AMaxImpl {
-    using CudaType = cuda_internal::DataType<T>;
-    __device__ CudaType Identity() { return cuda::NumericLimits<CudaType>::LowestOrInf(); }
-    __device__ CudaType MapIn(CudaType in, int64_t /*index*/) { return in; }
-    __device__ void Reduce(CudaType next, CudaType& accum) {
-        if (cuda::IsNan(next) || accum < next) {
-            accum = next;
-        }
-    }
-    __device__ CudaType MapOut(CudaType accum) { return accum; }
-};
-
-class CudaAMaxKernel : public AMaxKernel {
-public:
-    void Call(const Array& a, const Axes& axis, const Array& out) override {
-        Device& device = a.device();
-        CHAINERX_ASSERT(internal::IsValidReductionShape(a.shape(), axis, out.shape(), true));
-        device.CheckDevicesCompatible(a, out);
-        CudaSetDeviceScope scope{device.index()};
-        VisitDtype(out.dtype(), [&](auto pt) {
-            using T = typename decltype(pt)::type;
-            Reduce<T, T>(a, axis, out, AMaxImpl<T>{});
-        });
-    }
-};
-
-CHAINERX_CUDA_REGISTER_KERNEL(AMaxKernel, CudaAMaxKernel);
-
-template <typename T>
-struct AMinImpl {
-    using CudaType = cuda_internal::DataType<T>;
-    __device__ CudaType Identity() { return cuda::NumericLimits<CudaType>::MaxOrInf(); }
-    __device__ CudaType MapIn(CudaType in, int64_t /*index*/) { return in; }
-    __device__ void Reduce(CudaType next, CudaType& accum) {
-        if (cuda::IsNan(next) || accum > next) {
-            accum = next;
-        }
-    }
-    __device__ CudaType MapOut(CudaType accum) { return accum; }
-};
-
-class CudaAMinKernel : public AMinKernel {
-public:
-    void Call(const Array& a, const Axes& axis, const Array& out) override {
-        Device& device = a.device();
-        CHAINERX_ASSERT(internal::IsValidReductionShape(a.shape(), axis, out.shape(), true));
-        device.CheckDevicesCompatible(a, out);
-        CudaSetDeviceScope scope{device.index()};
-        VisitDtype(out.dtype(), [&](auto pt) {
-            using T = typename decltype(pt)::type;
-            Reduce<T, T>(a, axis, out, AMinImpl<T>{});
-        });
-    }
-};
-
-CHAINERX_CUDA_REGISTER_KERNEL(AMinKernel, CudaAMinKernel);
-
 }  // namespace
 }  // namespace cuda
 }  // namespace chainerx
