+++ chainerx_cc/chainerx/routines/math.cc
@@ -466,96 +466,6 @@ Array Sum(const Array& a, const OptionalAxes& axis, bool keepdims) {
     return out;
 }
 
-Array AMax(const Array& a, const OptionalAxes& axis, bool keepdims) {
-    Axes sorted_axis = internal::GetSortedAxesOrAll(axis, a.ndim());
-    Array out = internal::EmptyReduced(a.shape(), a.dtype(), sorted_axis, keepdims, a.device());
-
-    for (int8_t i : sorted_axis) {
-        if (a.shape()[i] == 0) {
-            throw DimensionError{"cannot compute the maximum along zero-sized axis"};
-        }
-    }
-
-    {
-        NoBackpropModeScope scope{};
-        a.device().backend().CallKernel<AMaxKernel>(a, sorted_axis, out);
-    }
-
-    BackwardBuilder bb{"amax", a, out};
-    if (BackwardBuilder::Target bt = bb.CreateTarget(0)) {
-        // a and out are used only for restoring the mask. We don't need graph nodes.
-        bt.Define([sorted_axis, a = a.AsGradStopped(), out = out.AsGradStopped(), keepdims](BackwardContext& bctx) {
-            const Array& gout = *bctx.output_grad();
-            CHAINERX_ASSERT(std::is_sorted(sorted_axis.begin(), sorted_axis.end()));
-
-            Array reshaped_gout{};
-            Array reshaped_out{};
-            if (keepdims) {
-                reshaped_gout = gout;
-                reshaped_out = out;
-            } else {
-                // Add broadcastable dimensions to out and gout
-                // for each one that was reduced in the forward operation
-                Shape shape = internal::ReduceShape(a.shape(), sorted_axis, true);
-                reshaped_gout = gout.Reshape(shape);
-                reshaped_out = out.Reshape(shape);
-            }
-
-            // Compute the gradient
-            // TODO(sonots): Use `where` if it becomes available.
-            Array cond = (a == reshaped_out).AsType(gout.dtype(), false);
-            bctx.input_grad() = reshaped_gout * cond;
-        });
-    }
-    bb.Finalize();
-    return out;
-}
-
-Array AMin(const Array& a, const OptionalAxes& axis, bool keepdims) {
-    Axes sorted_axis = internal::GetSortedAxesOrAll(axis, a.ndim());
-    Array out = internal::EmptyReduced(a.shape(), a.dtype(), sorted_axis, keepdims, a.device());
-
-    for (int8_t i : sorted_axis) {
-        if (a.shape()[i] == 0) {
-            throw DimensionError{"cannot compute the minimum along zero-sized axis"};
-        }
-    }
-
-    {
-        NoBackpropModeScope scope{};
-        a.device().backend().CallKernel<AMinKernel>(a, sorted_axis, out);
-    }
-
-    BackwardBuilder bb{"amin", a, out};
-    if (BackwardBuilder::Target bt = bb.CreateTarget(0)) {
-        // a and out are used only for restoring the mask. We don't need graph nodes.
-        bt.Define([sorted_axis, a = a.AsGradStopped(), out = out.AsGradStopped(), keepdims](BackwardContext& bctx) {
-            const Array& gout = *bctx.output_grad();
-            CHAINERX_ASSERT(std::is_sorted(sorted_axis.begin(), sorted_axis.end()));
-
-            Array reshaped_gout{};
-            Array reshaped_out{};
-            if (keepdims) {
-                reshaped_gout = gout;
-                reshaped_out = out;
-            } else {
-                // Add broadcastable dimensions to out and gout
-                // for each one that was reduced in the forward operation
-                Shape shape = internal::ReduceShape(a.shape(), sorted_axis, true);
-                reshaped_gout = gout.Reshape(shape);
-                reshaped_out = out.Reshape(shape);
-            }
-
-            // Compute the gradient
-            // TODO(sonots): Use `where` if it becomes available.
-            Array cond = (a == reshaped_out).AsType(gout.dtype(), false);
-            bctx.input_grad() = reshaped_gout * cond;
-        });
-    }
-    bb.Finalize();
-    return out;
-}
-
 namespace {
 
 // Calculates: x1 < x2 ? pos : neg
